requirements.txt
behave-django==1.1.0
dj-database-url==0.5.0
django==2.1.5
django-allauth==0.38.0
django-cors-middleware==1.3.1
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
Faker==1.0.0
gunicorn==19.9.0
Pillow==5.3.0
psycopg2-binary==2.7.6.1
ptvsd==4.2.0
pytz==2018.5
whitenoise==4.0